pyproject.toml
[project]
name = "dji-thermal"
version = "1.7.0"
description = "DJI热红外处理工具"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "numpy>=1.22.0",
    "pillow>=9.0.0",
    "tqdm>=4.65.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false